+++ src/clients/browser/characters.js
@@ -1,0 +1,8 @@
+import { characterAppWatcher, sidebarPortalWatcher } from "~/characters";
+
+const main = () => {
+  characterAppWatcher();
+  sidebarPortalWatcher();
+};
+
+main();