+++ src/clients/browser/monsters.js
@@ -1,0 +1,3 @@
+import { monsterWatcher } from "~/monsters";
+
+monsterWatcher();